add `ui/time-duration` tests
